requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
argparse==1.4.0
uuid==1.30
pathlib==1.0.1